Create README.md
The data set is a real time and it is aimed to solve some problem related to thief. R is used to pre-process and manipulate the data and a simple Bayes rule is used to find some solution to it. Using the formula the top 20 suspects are given from the hundreds of visitors which could have the owner reduce such theft in future or help police to investigate the problem.